server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Cyrus Attoun <jattoun1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.12"
uvicorn = "^0.34.3"
lancedb = "^0.17.0"
python-dotenv = "^1.0.0"
pillow = "^10.0.0"
python-multipart = "^0.0.20"
pandas = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
